pyproject.toml
[project]
name = "jubensha-2"
version = "0.1.0"
description = "AI剧本杀游戏demo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "dashscope>=1.23.8",
    # 数据库相关
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    # 存储相关
    "minio>=7.2.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "aiohttp>=3.12.13",
    "alembic>=1.16.4",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
]

[mypy]
disable-error-code = "assignment, arg-type"  # 忽略赋值

[dependency-groups]
dev = [
    "mypy>=1.17.0",
]
